pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zenrows-deviceprofiles"
version = "0.0.1"
description = "ZenRows Device Profiles API"
authors = [
  { name = "ZenRows" }
]
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "sqlalchemy>=2.0.30",
  "alembic>=1.13.1",
  "psycopg[binary]>=3.2",
  "httpx>=0.27.0",
  "boto3>=1.34.0",
  "aws-xray-sdk>=2.14.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=4.1.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["app"]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["tests*", "migrations*", "infra*", "ops*", "DOCS*", "scripts*"]
